refactor: Console class util 패키지에서 view 패키지로 이동
